Made improvements: - Consistency - Descriptive Exception - Renamed package - Removed redundancies
